Light reflections started!!!
I started to add the light reflection puzzle stuff! so far I have the emitter, reciever, and have a bouncer in the works! Currently, the bouncer is a bit janky, so it doesn't bounce in the correct angle, let alone render the mesh properly. Oh well, another problem for tomorrow!